# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: OpenBox API
  version: v1
  description: Versioned, owner-scoped API for the OpenBox instance engine.
servers:
  - url: http://127.0.0.1:8443
    description: Loopback-only default for local v0.1 deployments.
  - url: https://127.0.0.1:8443
    description: Optional TLS listener when an operator supplies a certificate and key.
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /v1/tokens:
    parameters:
      - { $ref: "#/components/parameters/APIVersion" }
    get:
      operationId: listTokens
      responses:
        "200": { description: Token metadata without secrets or hashes., content: { application/json: { schema: { $ref: "#/components/schemas/ListTokensResponse" } } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createToken
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateTokenRequest" } } }
      responses:
        "201": { description: Token secret, returned once., content: { application/json: { schema: { $ref: "#/components/schemas/CreatedToken" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/tokens/{token_id}:
    parameters:
      - { $ref: "#/components/parameters/APIVersion" }
      - { name: token_id, in: path, required: true, schema: { type: string } }
    delete:
      operationId: revokeToken
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses:
        "204": { description: Token revoked immediately. }
        default: { $ref: "#/components/responses/Error" }
  /v1/ssh-keys:
    parameters:
      - { $ref: "#/components/parameters/APIVersion" }
    get:
      operationId: listSSHKeys
      responses:
        "200": { description: Authorized SSH keys., content: { application/json: { schema: { $ref: "#/components/schemas/ListSSHKeysResponse" } } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createSSHKey
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateSSHKeyRequest" } } }
      responses:
        "201": { description: Normalized SSH key., content: { application/json: { schema: { $ref: "#/components/schemas/SSHKey" } } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/ssh-keys/{key_id}:
    parameters:
      - { $ref: "#/components/parameters/APIVersion" }
      - { name: key_id, in: path, required: true, schema: { type: string } }
    patch:
      operationId: updateSSHKey
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateSSHKeyRequest" } } }
      responses:
        "200": { description: SSH key label updated without changing key material., content: { application/json: { schema: { $ref: "#/components/schemas/SSHKey" } } } }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteSSHKey
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses:
        "204": { description: SSH key deleted. }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: openbox_session
      description: HttpOnly SameSite browser session. Unsafe requests also require X-CSRF-Token.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpenBox token
  headers:
    APIVersion:
      description: API contract version used for this response.
      schema: { type: string, enum: [v1] }
  parameters:
    APIVersion:
      name: X-OpenBox-API-Version
      in: header
      required: false
      description: Optional compatibility assertion. If present, it must be v1.
      schema: { type: string, enum: [v1] }
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for unsafe requests authenticated by the session cookie; ignored for bearer authentication.
      schema: { type: string, minLength: 20 }
  responses:
    Error:
      description: Stable error envelope.
      headers: { X-OpenBox-API-Version: { $ref: "#/components/headers/APIVersion" } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    TokenMetadata:
      type: object
      required: [id, name, scopes, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        scopes: { type: array, items: { type: string, enum: [owner] } }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
    CreatedToken:
      allOf:
        - { $ref: "#/components/schemas/TokenMetadata" }
        - type: object
          required: [secret]
          properties: { secret: { type: string, readOnly: true } }
    CreateTokenRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        scopes: { type: array, items: { type: string, enum: [owner] } }
        expires_at: { type: string, format: date-time }
    ListTokensResponse:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: "#/components/schemas/TokenMetadata" } } }
    SSHKey:
      type: object
      required: [id, label, fingerprint, public_key, created_at]
      properties:
        id: { type: string }
        label: { type: string }
        fingerprint: { type: string }
        public_key: { type: string }
        created_at: { type: string, format: date-time }
    CreateSSHKeyRequest:
      type: object
      additionalProperties: false
      required: [label, public_key]
      properties:
        label: { type: string, minLength: 1, maxLength: 100 }
        public_key: { type: string, minLength: 1, maxLength: 16384 }
    UpdateSSHKeyRequest:
      type: object
      additionalProperties: false
      required: [label]
      properties: { label: { type: string, minLength: 1, maxLength: 100 } }
    ListSSHKeysResponse:
      type: object
      required: [items]
      properties: { items: { type: array, items: { $ref: "#/components/schemas/SSHKey" } } }
...
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, retryable, request_id]
          additionalProperties: false
          properties:
            code: { type: string }
            message: { type: string }
            field: { type: string }
            retryable: { type: boolean }
            request_id: { type: string }